Sieve cubes of small primes.

+++ Decompose.Numerics/ArithmeticFunctions/DivisionFreeDivisorSummatoryFunction.cs
@@ -538,7 +538,7 @@ namespace Decompose.Numerics
                 Debug.Assert(eps == n % x);
                 Debug.Assert((beta & 3) == (n / x + 1) % 4);
                 Debug.Assert(delta == n / x - n / (x + 2));
-                Debug.Assert(gamma == 2 * (n / x) - (UInt128)(x - 2) * (UInt128)delta);
+                Debug.Assert(gamma == 2 * (n / x) - (BigInteger)(x - 2) * delta);
 
                 s ^= beta;
                 x -= 2;
